requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=14.0.0
Pillow>=10.0.0
PyYAML>=6.0
tqdm>=4.65.0

# Data downloading
img2dataset>=1.41.0

# Models (HuggingFace)
transformers>=4.36.0
diffusers>=0.25.0
accelerate>=0.25.0
safetensors>=0.4.0

# Filtering
imagehash>=4.3.1

# Clustering
scikit-learn>=1.3.0

# Sharding
webdataset>=0.2.86

# Visualization
matplotlib>=3.8.0

# CLIP
ftfy>=6.1.0
regex>=2023.0
